[refactor] App: relocate usecases and utils
