+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/data/repositories/TimelineRepositoryImpl.kt
@@ -8,13 +8,13 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.json.Json
-import uk.ryanwong.gmap2ics.app.utils.timezonemap.TimeZoneMapWrapper
 import uk.ryanwong.gmap2ics.data.datasources.local.LocalDataSource
 import uk.ryanwong.gmap2ics.data.datasources.local.impl.LocalDataSourceImpl
 import uk.ryanwong.gmap2ics.data.except
 import uk.ryanwong.gmap2ics.data.models.timeline.TimelineObjects
 import uk.ryanwong.gmap2ics.domain.models.timeline.Timeline
 import uk.ryanwong.gmap2ics.domain.repositories.TimelineRepository
+import uk.ryanwong.gmap2ics.domain.utils.timezonemap.TimeZoneMapWrapper
 import kotlin.coroutines.cancellation.CancellationException
 
 class TimelineRepositoryImpl(
