+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/di/UseCaseModule.kt
@@ -5,16 +5,16 @@
 package uk.ryanwong.gmap2ics.di
 
 import org.koin.dsl.module
-import uk.ryanwong.gmap2ics.domain.usecases.GetActivitySegmentVEventUseCase
-import uk.ryanwong.gmap2ics.domain.usecases.GetOutputFilenameUseCase
-import uk.ryanwong.gmap2ics.domain.usecases.GetPlaceVisitVEventUseCase
-import uk.ryanwong.gmap2ics.domain.usecases.VEventFromActivitySegmentUseCase
-import uk.ryanwong.gmap2ics.usecases.GetActivitySegmentVEventUseCaseImpl
-import uk.ryanwong.gmap2ics.usecases.GetOutputFilenameUseCaseImpl
-import uk.ryanwong.gmap2ics.usecases.GetPlaceVisitVEventUseCaseImpl
-import uk.ryanwong.gmap2ics.usecases.VEventFromActivitySegmentUseCaseImpl
-import uk.ryanwong.gmap2ics.usecases.VEventFromChildVisitUseCaseImpl
-import uk.ryanwong.gmap2ics.usecases.VEventFromPlaceVisitUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.GetActivitySegmentVEventUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.GetOutputFilenameUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.GetPlaceVisitVEventUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.VEventFromActivitySegmentUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.VEventFromChildVisitUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.VEventFromPlaceVisitUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.GetActivitySegmentVEventUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.GetOutputFilenameUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.GetPlaceVisitVEventUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.VEventFromActivitySegmentUseCase
 
 val useCaseModule = module {
     factory<GetPlaceVisitVEventUseCase> {
