+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/models/timeline/placevisit/ChildVisitMapper.kt
@@ -4,9 +4,9 @@
 
 package uk.ryanwong.gmap2ics.domain.models.timeline.placevisit
 
-import uk.ryanwong.gmap2ics.app.utils.timezonemap.TimeZoneMapWrapper
 import uk.ryanwong.gmap2ics.domain.models.RawTimestamp
 import uk.ryanwong.gmap2ics.domain.models.timeline.toDomainModel
+import uk.ryanwong.gmap2ics.domain.utils.timezonemap.TimeZoneMapWrapper
 
 fun uk.ryanwong.gmap2ics.data.models.timeline.ChildVisit.toDomainModel(timeZoneMap: TimeZoneMapWrapper): ChildVisit? {
     //  If a child visit does not have a valid duration start & end, we simply drop it during conversion
