+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/GetActivitySegmentVEventUseCaseImpl.kt
@@ -2,13 +2,13 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.usecases
+package uk.ryanwong.gmap2ics.domain.usecases
 
 import uk.ryanwong.gmap2ics.domain.models.ActivityType
 import uk.ryanwong.gmap2ics.domain.models.VEvent
 import uk.ryanwong.gmap2ics.domain.models.timeline.activity.ActivitySegment
-import uk.ryanwong.gmap2ics.domain.usecases.GetActivitySegmentVEventUseCase
-import uk.ryanwong.gmap2ics.domain.usecases.VEventFromActivitySegmentUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.GetActivitySegmentVEventUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.VEventFromActivitySegmentUseCase
 
 class GetActivitySegmentVEventUseCaseImpl(
     private val vEventFromActivitySegmentUseCase: VEventFromActivitySegmentUseCase,
