--- composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/VEventFromPlaceVisitUseCase.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
- */
-
-package uk.ryanwong.gmap2ics.domain.usecases
-
-import uk.ryanwong.gmap2ics.domain.models.VEvent
-import uk.ryanwong.gmap2ics.domain.models.timeline.placevisit.PlaceVisit
-
-interface VEventFromPlaceVisitUseCase {
-    suspend operator fun invoke(
-        placeVisit: PlaceVisit,
-        enablePlacesApiLookup: Boolean,
-    ): VEvent
-}
