+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/VEventFromPlaceVisitUseCaseImpl.kt
@@ -2,13 +2,13 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.usecases
+package uk.ryanwong.gmap2ics.domain.usecases
 
 import io.github.aakira.napier.Napier
 import uk.ryanwong.gmap2ics.domain.models.VEvent
 import uk.ryanwong.gmap2ics.domain.models.timeline.placevisit.PlaceVisit
 import uk.ryanwong.gmap2ics.domain.repositories.PlaceDetailsRepository
-import uk.ryanwong.gmap2ics.domain.usecases.VEventFromPlaceVisitUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.VEventFromPlaceVisitUseCase
 
 class VEventFromPlaceVisitUseCaseImpl(
     private val placeDetailsRepository: PlaceDetailsRepository,
