+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/interfaces/GetPlaceVisitVEventUseCase.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.domain.usecases
+package uk.ryanwong.gmap2ics.domain.usecases.interfaces
 
 import kotlinx.coroutines.flow.SharedFlow
 import uk.ryanwong.gmap2ics.domain.models.UILogEntry
