+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/interfaces/VEventFromActivitySegmentUseCase.kt
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
+ */
+
+package uk.ryanwong.gmap2ics.domain.usecases.interfaces
+
+import uk.ryanwong.gmap2ics.domain.models.VEvent
+import uk.ryanwong.gmap2ics.domain.models.timeline.activity.ActivitySegment
+
+interface VEventFromActivitySegmentUseCase {
+    suspend operator fun invoke(
+        activitySegment: ActivitySegment,
+        enablePlacesApiLookup: Boolean,
+    ): VEvent
+}
