+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/domain/usecases/interfaces/VEventFromPlaceVisitUseCase.kt
@@ -2,14 +2,14 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.domain.usecases
+package uk.ryanwong.gmap2ics.domain.usecases.interfaces
 
 import uk.ryanwong.gmap2ics.domain.models.VEvent
-import uk.ryanwong.gmap2ics.domain.models.timeline.activity.ActivitySegment
+import uk.ryanwong.gmap2ics.domain.models.timeline.placevisit.PlaceVisit
 
-interface VEventFromActivitySegmentUseCase {
+interface VEventFromPlaceVisitUseCase {
     suspend operator fun invoke(
-        activitySegment: ActivitySegment,
+        placeVisit: PlaceVisit,
         enablePlacesApiLookup: Boolean,
     ): VEvent
 }
