+++ composeApp/src/commonTest/kotlin/uk/ryanwong/gmap2ics/usecases/fakes/FakeGetActivitySegmentVEventUseCase.kt
@@ -7,7 +7,7 @@ package uk.ryanwong.gmap2ics.usecases.fakes
 import uk.ryanwong.gmap2ics.domain.models.ActivityType
 import uk.ryanwong.gmap2ics.domain.models.VEvent
 import uk.ryanwong.gmap2ics.domain.models.timeline.activity.ActivitySegment
-import uk.ryanwong.gmap2ics.domain.usecases.GetActivitySegmentVEventUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.GetActivitySegmentVEventUseCase
 
 class FakeGetActivitySegmentVEventUseCase : GetActivitySegmentVEventUseCase {
     var useCaseResponse: VEvent? = null
