+++ composeApp/src/commonTest/kotlin/uk/ryanwong/gmap2ics/usecases/fakes/FakeVEventFromActivitySegmentUseCase.kt
@@ -6,7 +6,7 @@ package uk.ryanwong.gmap2ics.usecases.fakes
 
 import uk.ryanwong.gmap2ics.domain.models.VEvent
 import uk.ryanwong.gmap2ics.domain.models.timeline.activity.ActivitySegment
-import uk.ryanwong.gmap2ics.domain.usecases.VEventFromActivitySegmentUseCase
+import uk.ryanwong.gmap2ics.domain.usecases.interfaces.VEventFromActivitySegmentUseCase
 
 class FakeVEventFromActivitySegmentUseCase : VEventFromActivitySegmentUseCase {
     var useCaseResponse: VEvent? = null
