+++ composeApp/src/commonTest/kotlin/uk/ryanwong/gmap2ics/usecases/impl/VEventFromPlaceVisitUseCaseImplTest.kt
@@ -16,7 +16,7 @@ import uk.ryanwong.gmap2ics.domain.models.timeline.LatLng
 import uk.ryanwong.gmap2ics.domain.models.timeline.Location
 import uk.ryanwong.gmap2ics.domain.models.timeline.PlaceDetails
 import uk.ryanwong.gmap2ics.domain.models.timeline.placevisit.PlaceVisit
-import uk.ryanwong.gmap2ics.usecases.VEventFromPlaceVisitUseCaseImpl
+import uk.ryanwong.gmap2ics.domain.usecases.VEventFromPlaceVisitUseCaseImpl
 import us.dustinj.timezonemap.TimeZone
 
 internal class VEventFromPlaceVisitUseCaseImplTest : FreeSpec() {
